change bottom buttons
